[improve] 컴파일 경고 3건 정리 #SW1-1829
clean 리빌드 시 vins_estimator 경고 3건을 제거(에러 0/경고 0 확인). 동작·수치 로직 불변.

- factor/wheel_factor.h: Jr_drdsw를 Identity로 명시 초기화.
  rightJacobianSO3가 내부에서 항상 채우지만 const-ref 우회 대입을 컴파일러가
  추적 못 해 -Wmaybe-uninitialized 오탐 발생 → 무해하게 차단(신규 휠 코드)
- feature_tracker.cpp: size_a에 [[maybe_unused]] (ROS_DEBUG 컷 시 unused 경고)
- feature_manager.cpp: no_depth_num을 size_t 캐스트해 부호 비교(-Wsign-compare) 정합

뒤 2건은 휠 작업과 무관한 기존 경고지만 같은 패키지 빌드 정리 차원에서 함께 수정.

+++ vins_estimator/src/feature_manager/feature_manager.cpp
@@ -462,7 +462,7 @@ void FeatureManager::triangulateWithDepth(Vector3d _Ps[], Vector3d _tic[], Matri
         {
             if (rough_depths.empty())
             {
-                if (no_depth_num == it_per_id.feature_per_frame.size())
+                if (static_cast<size_t>(no_depth_num) == it_per_id.feature_per_frame.size())  // int↔size_type 부호 비교 경고 차단
                 {
                     int imu_j = imu_i - 1;
